# public/api-docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: GhostClass API
  version: 1.3.3
  description: |
    **GhostClass API** provides endpoints for managing attendance synchronization with EzyGo.
    
    ## Authentication
    
    The API supports two authentication methods:
    
    1. **Bearer Token (CRON_SECRET)**: For automated cron jobs
       - Add `Authorization: Bearer <CRON_SECRET>` header
       - Used by automated sync processes
    
    2. **Session Cookie (SupabaseAuth)**: For authenticated users
       - Automatically included when logged in through the web interface
       - Managed by Supabase Auth
    
    ## Rate Limiting
    
    All endpoints are rate-limited to prevent abuse:
    
    - `/api/auth/save-token` & `/api/cron/sync`: Configurable via environment variables
    - Contact Form (Server Action): Configurable via environment variables
    
    Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets
    
    ## Base URLs
    
    - Production: Configured via `NEXT_PUBLIC_APP_URL`
    - Development: `http://localhost:3000`
    
    ## Error Handling
    
    All error responses follow a consistent format with appropriate HTTP status codes and descriptive messages.

  contact:
    name: GhostClass Support
    email: ${NEXT_PUBLIC_APP_EMAIL}
    url: ${NEXT_PUBLIC_GITHUB_URL}
  
  license:
...
servers:
...
paths:
...
  /api/cron/sync:
    get:
      tags:
        - Sync
      summary: Sync attendance data with EzyGo
      description: |
        Synchronizes attendance records from EzyGo with the local database. Supports both automated cron jobs and authenticated user-initiated syncs.
        
        **What it does:**
        - Fetches latest attendance data from EzyGo API
        - Compares with local tracking records
        - Detects and resolves conflicts (course mismatches, attendance discrepancies)
        - Sends email notifications for conflicts
        - Updates sync timestamps
        
        **Authentication:** 
        - **Option 1 (Cron)**: Bearer token with CRON_SECRET in Authorization header
        - **Option 2 (Users)**: Valid session cookie (authenticated users)
        
        **Rate Limiting:** Configured via RATE_LIMIT_REQUESTS and RATE_LIMIT_WINDOW environment variables
        
        **Side Effects:**
        - Deletes verified/matched attendance records
        - Updates conflicted records with 'correction' status
        - Creates notifications in the database
        - Sends email alerts for attendance conflicts
        - Updates `last_synced_at` timestamp for users
        
        **Use Cases:**
        - Automated daily/hourly sync via cron job
        - Manual sync triggered by users from dashboard
        - Batch sync for multiple users (cron only)
        - Single user sync with optional username filter
      
      operationId: syncAttendance
      
      security:
        - BearerAuth: []
        - SupabaseAuth: []
      
      parameters:
        - name: username
          in: query
          required: false
          description: |
            Filter sync to a specific user by username. Only applicable when using Bearer token authentication (cron mode).
            
            - If provided with cron auth: syncs only the specified user
            - If omitted with cron auth: syncs up to 10 least-recently-synced users
            - Ignored for authenticated user requests (always syncs the current user)
          schema:
            type: string
            example: "student123"
      
      responses:
        '200':
          description: Sync completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Whether sync operation completed successfully
                    example: true
                  processed:
                    type: integer
                    description: Number of users processed
                    example: 5
                  deletions:
                    type: integer
                    description: Number of tracking records deleted (verified/matched)
                    example: 23
                  conflicts:
                    type: integer
                    description: Number of attendance conflicts detected
                    example: 2
                  updates:
                    type: integer
                    description: Number of records updated to 'correction' status
                    example: 2
                  errors:
                    type: integer
                    description: Number of users that failed to sync
                    example: 0
              examples:
                successfulSync:
                  summary: Successful sync with multiple users
                  value:
                    success: true
                    processed: 5
                    deletions: 23
                    conflicts: 2
                    updates: 2
                    errors: 0
                noUsers:
                  summary: No eligible users to sync
                  value:
                    success: true
                    processed: 0
                    deletions: 0
                    conflicts: 0
                    updates: 0
                    errors: 0
        
        '401':
          description: Unauthorized - Invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                unauthorized:
                  summary: Not authenticated
                  value:
                    message: "Unauthorized"
        
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Reset:
              schema:
                type: integer
              description: Unix timestamp when the rate limit resets
              example: 1706108460
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
              examples:
                tooManyRequests:
                  summary: Rate limit exceeded
                  value:
                    error: "Too many requests"
                    retryAfter: 60
        
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                serverError:
                  summary: Configuration error
                  value:
                    error: "Server configuration error: NEXT_PUBLIC_APP_URL is not set"
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message describing what went wrong
        message:
          type: string
          description: Detailed error message
      example:
        error: "Internal Server Error"
...
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          description: Rate limit error message
          example: "Too many requests. Slow down!"
        retryAfter:
          type: integer
          description: Unix timestamp (in seconds) when the rate limit resets. Matches the `X-RateLimit-Reset` header.
          example: 1735689600
      required:
        - error
        - retryAfter
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token authentication using CRON_SECRET environment variable.
        Used for automated cron jobs.
        
        Example: `Authorization: Bearer <CRON_SECRET>`
    
    SupabaseAuth:
      type: apiKey
      in: cookie
      name: sb-*
      description: |
        Session cookie authentication managed by Supabase Auth.
        Automatically included when users are logged in through the web interface.
        Cookie names typically start with 'sb-' prefix.
...